+++ packages/core/src/agents/mcp-agent.ts
@@ -1,18 +1,22 @@
-import { Client } from "@modelcontextprotocol/sdk/client/index.js";
+import { Client, type ClientOptions } from "@modelcontextprotocol/sdk/client/index.js";
 import { SSEClientTransport } from "@modelcontextprotocol/sdk/client/sse.js";
 import {
   StdioClientTransport,
   type StdioServerParameters,
   getDefaultEnvironment,
 } from "@modelcontextprotocol/sdk/client/stdio.js";
+import type { RequestOptions } from "@modelcontextprotocol/sdk/shared/protocol.js";
 import type { Transport } from "@modelcontextprotocol/sdk/shared/transport.js";
 import { UriTemplate } from "@modelcontextprotocol/sdk/shared/uriTemplate.js";
 import type {
   CallToolResult,
   GetPromptResult,
+  Implementation,
   ReadResourceResult,
+  Request,
 } from "@modelcontextprotocol/sdk/types.js";
-import {} from "zod";
+import pRetry from "p-retry";
+import { type ZodType, z } from "zod";
 import type { Context } from "../execution-engine/context.js";
 import { logger } from "../utils/logger.js";
 import {
@@ -20,11 +24,12 @@ import {
   resourceFromMCPResource,
   toolFromMCPTool,
 } from "../utils/mcp-utils.js";
-import { createAccessorArray } from "../utils/type-utils.js";
+import { type PromiseOrValue, checkArguments, createAccessorArray } from "../utils/type-utils.js";
 import { Agent, type AgentOptions, type Message } from "./agent.js";
 
 const MCP_AGENT_CLIENT_NAME = "MCPAgent";
 const MCP_AGENT_CLIENT_VERSION = "0.0.1";
+const DEFAULT_MAX_RECONNECTS = 10;
 
 const debug = logger.base.extend("mcp");
 
@@ -40,6 +45,16 @@ export type MCPServerOptions = SSEServerParameters | StdioServerParameters;
 
 export type SSEServerParameters = {
   url: string;
+  /**
+   * Whether to automatically reconnect to the server if the connection is lost.
+   * @default 10 set to 0 to disable automatic reconnection
+   */
+  maxReconnects?: number;
+  /**
+   * A function that determines whether to reconnect to the server based on the error.
+   * default to reconnect on all errors.
+   */
+  shouldReconnect?: (error: Error) => boolean;
 };
 
 function isSSEServerParameters(
@@ -70,20 +85,23 @@ export class MCPAgent extends Agent {
   static from(options: MCPServerOptions): Promise<MCPAgent>;
   static from(options: MCPAgentOptions): MCPAgent;
   static from(options: MCPAgentOptions | MCPServerOptions): MCPAgent | Promise<MCPAgent> {
+    checkArguments("MCPAgent.from", mcpAgentOptionsSchema, options);
+
     if (isSSEServerParameters(options)) {
-      const transport = new SSEClientTransport(new URL(options.url));
+      const transport = () => new SSEClientTransport(new URL(options.url));
       return MCPAgent.fromTransport(transport, options);
     }
 
     if (isStdioServerParameters(options)) {
-      const transport = new StdioClientTransport({
-        ...options,
-        env: {
-          ...getDefaultEnvironment(),
-          ...options.env,
-        },
-        stderr: "pipe",
-      });
+      const transport = () =>
+        new StdioClientTransport({
+          ...options,
+          env: {
+            ...getDefaultEnvironment(),
+            ...options.env,
+          },
+          stderr: "pipe",
+        });
       return MCPAgent.fromTransport(transport, options);
     }
 
@@ -91,13 +109,19 @@ export class MCPAgent extends Agent {
   }
 
   private static async fromTransport(
-    transport: Transport,
+    transportCreator: () => Transport,
     options: MCPAgentOptions | MCPServerOptions,
   ): Promise<MCPAgent> {
-    const client = new Client({
-      name: MCP_AGENT_CLIENT_NAME,
-      version: MCP_AGENT_CLIENT_VERSION,
-    });
+    const client = new ClientWithReconnect(
+      {
+        name: MCP_AGENT_CLIENT_NAME,
+        version: MCP_AGENT_CLIENT_VERSION,
+      },
+      undefined,
+      isSSEServerParameters(options) ? { transportCreator, ...options } : undefined,
+    );
+
+    const transport = transportCreator();
 
     await debug.spinner(
       client.connect(transport),
@@ -117,7 +141,7 @@ export class MCPAgent extends Agent {
           .spinner(client.listTools(), `Listing tools from ${mcpServer}`, ({ tools }) =>
             debug("%O", tools),
           )
-          .then(({ tools }) => tools.map((tool) => toolFromMCPTool(client, tool)))
+          .then(({ tools }) => tools.map((tool) => toolFromMCPTool(tool, { client })))
       : undefined;
 
     const prompts = isPromptsAvailable
@@ -125,7 +149,7 @@ export class MCPAgent extends Agent {
           .spinner(client.listPrompts(), `Listing prompts from ${mcpServer}`, ({ prompts }) =>
             debug("%O", prompts),
           )
-          .then(({ prompts }) => prompts.map((prompt) => promptFromMCPPrompt(client, prompt)))
+          .then(({ prompts }) => prompts.map((prompt) => promptFromMCPPrompt(prompt, { client })))
       : undefined;
 
     const resources = isResourcesAvailable
@@ -143,7 +167,7 @@ export class MCPAgent extends Agent {
           )
           .then(([{ resources }, { resourceTemplates }]) =>
             [...resources, ...resourceTemplates].map((resource) =>
-              resourceFromMCPResource(client, resource),
+              resourceFromMCPResource(resource, { client }),
             ),
           )
       : undefined;
@@ -189,18 +213,77 @@ export class MCPAgent extends Agent {
   }
 }
 
-export interface MCPToolBaseOptions<I extends Message, O extends Message>
+export interface ClientWithReconnectOptions {
+  transportCreator?: () => PromiseOrValue<Transport>;
+  maxReconnects?: number;
+  shouldReconnect?: (error: Error) => boolean;
+}
+
+class ClientWithReconnect extends Client {
+  constructor(
+    info: Implementation,
+    options?: ClientOptions,
+    private reconnectOptions?: ClientWithReconnectOptions,
+  ) {
+    super(info, options);
+  }
+
+  private shouldReconnect(error: Error): boolean {
+    const { transportCreator, shouldReconnect, maxReconnects } = this.reconnectOptions || {};
+
+    if (!transportCreator || maxReconnects === 0) return false;
+    if (!shouldReconnect) return true; // default to reconnect on all errors
+
+    return shouldReconnect(error);
+  }
+
+  private async reconnect() {
+    const transportCreator = this.reconnectOptions?.transportCreator;
+    if (!transportCreator) throw new Error("reconnect requires a transportCreator");
+
+    await pRetry(
+      async () => {
+        await this.close();
+        await this.connect(await transportCreator());
+      },
+      {
+        retries: this.reconnectOptions?.maxReconnects ?? DEFAULT_MAX_RECONNECTS,
+        shouldRetry: this.shouldReconnect,
+        onFailedAttempt: (error) => debug("Reconnect attempt failed: %O", error),
+      },
+    );
+  }
+
+  override async request<T extends ZodType<object>>(
+    request: Request,
+    resultSchema: T,
+    options?: RequestOptions,
+  ): Promise<z.infer<T>> {
+    try {
+      return await super.request(request, resultSchema, options);
+    } catch (error) {
+      if (this.shouldReconnect(error)) {
+        debug("Error occurred, reconnecting to MCP server: %O", error);
+        await this.reconnect();
+        return await super.request(request, resultSchema, options);
+      }
+      throw error;
+    }
+  }
+}
+
+export interface MCPBaseOptions<I extends Message = Message, O extends Message = Message>
   extends AgentOptions<I, O> {
-  client: Client;
+  client: ClientWithReconnect;
 }
 
 export abstract class MCPBase<I extends Message, O extends Message> extends Agent<I, O> {
-  constructor(options: MCPToolBaseOptions<I, O>) {
+  constructor(options: MCPBaseOptions<I, O>) {
     super(options);
     this.client = options.client;
   }
 
-  protected client: Client;
+  protected client: ClientWithReconnect;
 
   protected get mcpServer() {
     return getMCPServerName(this.client);
@@ -235,7 +318,7 @@ export class MCPPrompt extends MCPBase<MCPPromptInput, GetPromptResult> {
   }
 }
 
-export interface MCPResourceOptions extends MCPToolBaseOptions<MCPPromptInput, ReadResourceResult> {
+export interface MCPResourceOptions extends MCPBaseOptions<MCPPromptInput, ReadResourceResult> {
   uri: string;
 }
 
@@ -267,3 +350,21 @@ function getMCPServerName(client: Client): string | undefined {
   const { name, version } = info;
   return `${name}@${version}`;
 }
+
+const mcpAgentOptionsSchema: ZodType<
+  MCPAgentOptions | SSEServerParameters | StdioServerParameters
+> = z.union([
+  z.object({
+    client: z.instanceof(Client),
+    prompts: z.array(z.instanceof(MCPPrompt)).optional(),
+    resources: z.array(z.instanceof(MCPResource)).optional(),
+  }),
+  z.object({
+    url: z.string(),
+  }),
+  z.object({
+    command: z.string(),
+    args: z.array(z.string()).optional(),
+    env: z.record(z.string()).optional(),
+  }),
+]);