+++ packages/core/src/execution-engine/execution-engine.ts
@@ -1,44 +1,75 @@
 import EventEmitter from "node:events";
-import { isNil } from "lodash-es";
+import { z } from "zod";
 import { Agent, type Message } from "../agents/agent.js";
-import { isTransferAgentOutput, transferAgentOutputKey } from "../agents/types.js";
-import { UserAgent } from "../agents/user-agent.js";
-import type { ChatModel } from "../models/chat-model.js";
-import { createMessage } from "../prompt/prompt-builder.js";
-import type { Context, Runnable } from "./context.js";
+import {} from "../agents/types.js";
+import type { UserAgent } from "../agents/user-agent.js";
+import { load } from "../loader/index.js";
+import { ChatModel } from "../models/chat-model.js";
+import { checkArguments, createAccessorArray } from "../utils/type-utils.js";
+import { type ContextLimits, ExecutionContext, type Runnable } from "./context.js";
 import {
   type MessagePayload,
   MessageQueue,
   type MessageQueueListener,
-  type MessageRequest,
   type Unsubscribe,
 } from "./message-queue.js";
 
 export interface ExecutionEngineOptions {
+  name?: string;
+  description?: string;
   model?: ChatModel;
   tools?: Agent[];
   agents?: Agent[];
+  limits?: ContextLimits;
 }
 
-export class ExecutionEngine extends EventEmitter implements Context {
+export class ExecutionEngine extends EventEmitter {
+  static async load({
+    path,
+    ...options
+  }: { path: string } & ExecutionEngineOptions): Promise<ExecutionEngine> {
+    const { model, agents, tools, ...aigne } = await load({ path });
+    return new ExecutionEngine({
+      model,
+      ...options,
+      name: options.name || aigne.name || undefined,
+      description: options.description || aigne.description || undefined,
+      agents: agents.concat(options.agents ?? []),
+      tools: tools.concat(options.tools ?? []),
+    });
+  }
+
   constructor(options?: ExecutionEngineOptions) {
+    if (options) checkArguments("ExecutionEngine", executionEngineOptionsSchema, options);
+
     super();
+    this.name = options?.name;
+    this.description = options?.description;
     this.model = options?.model;
-    this.tools = options?.tools ?? [];
+    this.limits = options?.limits;
+    if (options?.tools?.length) this.tools.push(...options.tools);
     if (options?.agents?.length) this.addAgent(...options.agents);
 
     this.initProcessExitHandler();
   }
 
+  name?: string;
+
+  description?: string;
+
   readonly messageQueue = new MessageQueue();
 
   model?: ChatModel;
 
-  tools: Agent[];
+  readonly tools = createAccessorArray<Agent>([], (arr, name) => arr.find((i) => i.name === name));
 
-  private agents: Agent[] = [];
+  readonly agents = createAccessorArray<Agent>([], (arr, name) => arr.find((i) => i.name === name));
+
+  limits?: ContextLimits;
 
   addAgent(...agents: Agent[]) {
+    checkArguments("ExecutionEngine.addAgent", executionEngineAddAgentArgsSchema, agents);
+
     for (const agent of agents) {
       this.agents.push(agent);
 
@@ -46,8 +77,8 @@ export class ExecutionEngine extends EventEmitter implements Context {
     }
   }
 
-  private createMessageFromInput(message: Message | string): Message {
-    return typeof message === "string" ? createMessage(message) : message;
+  newContext() {
+    return new ExecutionContext(this);
   }
 
   /**
@@ -57,13 +88,7 @@ export class ExecutionEngine extends EventEmitter implements Context {
    * @param from the agent who publish the message, if not provided, it will be treated as a user message
    */
   publish(topic: string | string[], message: Message | string, from?: Agent) {
-    const request: MessageRequest = {
-      role: !from || from instanceof UserAgent ? "user" : "agent",
-      source: from?.name,
-      message: this.createMessageFromInput(message),
-    };
-
-    this.messageQueue.publish(topic, request);
+    return new ExecutionContext(this).publish(topic, message, from);
   }
 
   /**
@@ -92,105 +117,40 @@ export class ExecutionEngine extends EventEmitter implements Context {
   call<I extends Message, O extends Message>(
     agent: Runnable<I, O>,
     message: I | string,
-    options: { returnActiveAgent?: true },
+    options: { returnActiveAgent: true },
   ): Promise<[O, Runnable]>;
   call<I extends Message, O extends Message>(
     agent: Runnable<I, O>,
     message?: I | string,
     options?: { returnActiveAgent?: boolean },
-  ): UserAgent | Promise<O | [O, Runnable]> {
-    if (isNil(message)) {
-      let activeAgent: Runnable = agent;
-
-      return UserAgent.from({
-        context: this,
-        process: async (input, context) => {
-          const [output, agent] = await context.call(activeAgent, input, {
-            returnActiveAgent: true,
-          });
-          activeAgent = agent;
-          return output;
-        },
-      });
-    }
-
-    return this._call(agent, this.createMessageFromInput(message)).then((result) =>
-      options?.returnActiveAgent ? [result.output, result.agent] : result.output,
-    );
-  }
-
-  /**
-   * Call an agent with a message and return the output and the active agent
-   * @param agent Agent to call
-   * @param message Message to pass to the agent
-   * @returns the output of the agent and the final active agent
-   */
-  private async _call<I extends Message, O extends Message>(agent: Runnable<I, O>, message: I) {
-    const { output, agent: activeAgent } = await this.callAgent(agent, message);
-
-    if (activeAgent instanceof Agent) {
-      const publishTopics =
-        typeof activeAgent.publishTopic === "function"
-          ? await activeAgent.publishTopic(output)
-          : activeAgent.publishTopic;
-
-      if (publishTopics?.length) {
-        this.publish(publishTopics, output, activeAgent);
-      }
-    }
-
-    return { output, agent: activeAgent };
+  ): UserAgent<I, O> | Promise<O | [O, Runnable]>;
+  call<I extends Message, O extends Message>(
+    agent: Runnable<I, O>,
+    message?: I | string,
+    options?: { returnActiveAgent?: boolean },
+  ): UserAgent<I, O> | Promise<O | [O, Runnable]> {
+    return new ExecutionContext(this).call(agent, message, options);
   }
 
   subscribe(topic: string, listener?: undefined): Promise<MessagePayload>;
   subscribe(topic: string, listener: MessageQueueListener): Unsubscribe;
   subscribe(topic: string, listener?: MessageQueueListener): Unsubscribe | Promise<MessagePayload>;
   subscribe(topic: string, listener?: MessageQueueListener): Unsubscribe | Promise<MessagePayload> {
+    checkArguments("ExecutionEngine.subscribe", executionEngineSubscribeArgsSchema, {
+      topic,
+      listener,
+    });
+
     return this.messageQueue.subscribe(topic, listener);
   }
 
   unsubscribe(topic: string, listener: MessageQueueListener) {
+    checkArguments("ExecutionEngine.unsubscribe", executionEngineUnsubscribeArgsSchema, {
+      topic,
+      listener,
+    });
+
     this.messageQueue.unsubscribe(topic, listener);
-  }
-
-  private async callAgent<I extends Message, O extends Message>(
-    agent: Runnable<I, O>,
-    input: I,
-  ): Promise<{ agent: Runnable; output: O }> {
-    let activeAgent: Runnable = agent;
-    let output: O | undefined;
-
-    for (;;) {
-      let result: Message | Agent;
-
-      if (typeof activeAgent === "function") {
-        result = await activeAgent(input, this);
-      } else {
-        result = await activeAgent.call(input, this);
-      }
-
-      if (!(result instanceof Agent)) output = result as O;
-
-      const transferToAgent =
-        result instanceof Agent
-          ? result
-          : isTransferAgentOutput(result)
-            ? result[transferAgentOutputKey].agent
-            : undefined;
-
-      if (transferToAgent) {
-        activeAgent = transferToAgent;
-      } else {
-        break;
-      }
-    }
-
-    if (!output) throw new Error("Unexpected empty output");
-
-    return {
-      agent: activeAgent,
-      output,
-    };
   }
 
   async shutdown() {
@@ -208,3 +168,21 @@ export class ExecutionEngine extends EventEmitter implements Context {
     process.on("exit", shutdownAndExit);
   }
 }
+
+const executionEngineOptionsSchema = z.object({
+  model: z.instanceof(ChatModel).optional(),
+  tools: z.array(z.instanceof(Agent)).optional(),
+  agents: z.array(z.instanceof(Agent)).optional(),
+});
+
+const executionEngineAddAgentArgsSchema = z.array(z.instanceof(Agent));
+
+const executionEngineSubscribeArgsSchema = z.object({
+  topic: z.string(),
+  listener: z.function(z.tuple([z.any()]), z.any()).optional(),
+});
+
+const executionEngineUnsubscribeArgsSchema = z.object({
+  topic: z.string(),
+  listener: z.function(z.tuple([z.any()]), z.any()),
+});