+++ examples/workflow-concurrency/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.3.0](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-concurrency-v1.2.1...example-workflow-concurrency-v1.3.0) (2025-04-08)
+
+
+### Features
+
+* add agent library support ([#51](https://github.com/AIGNE-io/aigne-framework/issues/51)) ([1f0d34d](https://github.com/AIGNE-io/aigne-framework/commit/1f0d34ddda3154283a4bc958ddb9b68b4ac106b0))
+
+## [1.2.1](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-concurrency-v1.2.0...example-workflow-concurrency-v1.2.1) (2025-03-27)
+
+
+### Bug Fixes
+
+* **dx:** export models/utils in submodules ([#43](https://github.com/AIGNE-io/aigne-framework/issues/43)) ([bd561b3](https://github.com/AIGNE-io/aigne-framework/commit/bd561b397de816f04c2d63d58538e81fba82fc7f))
+
 ## [1.2.0](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-concurrency-v1.1.0...example-workflow-concurrency-v1.2.0) (2025-03-24)